consolidate events.go
